ai_interview_system/requirements.txt
fastapi
uvicorn
numpy
faster-whisper
sentence-transformers
librosa
groq
python-dotenv
deepface
opencv-python
mediapipe
tf-keras
